Shorten command name
This lets the user type in `pos`, as well as `marketplace-kit` when running commands.  To test, install package and then run npm link.  `pos` should then be available as a command in terminal and have all the same functionality that `marketplace-kit` does.

Not sure if you want this, but it seems to be more on brand and also a lot shorter/easier to type!